+++ src/source/RazorEngine.Core/Compilation/ICompilerService.cs
@@ -3,8 +3,7 @@
     using System;
     using System.Collections.Generic;
     using System.Reflection;
-
-    using Inspectors;
+    
     using RazorEngine.Compilation.ReferenceResolver;
     using System.Security;
 
@@ -14,13 +13,6 @@
     public interface ICompilerService
     {
         #region Properties
-#if !RAZOR4
-        /// <summary>
-        /// Gets or sets the set of code inspectors.
-        /// </summary>
-        [Obsolete("This API is obsolete and will be removed in the next version (Razor4 doesn't use CodeDom for code-generation)!")]
-        IEnumerable<ICodeInspector> CodeInspectors { get; set; }
-#endif
 
         /// <summary>
         /// Gets or sets the reference resolver.