+++ src/source/RazorEngine.Core/Configuration/Fluent/FluentTemplateServiceConfiguration.cs
@@ -5,7 +5,6 @@
     using System.Diagnostics.Contracts;
 
     using Compilation;
-    using Compilation.Inspectors;
     using Templating;
     using Text;
     using RazorEngine.Compilation.ReferenceResolver;
@@ -71,18 +70,7 @@
         {
             get { return _innerConfig.BaseTemplateType; }
         }
-        
-#if !RAZOR4
-        /// <summary>
-        /// Gets the set of code inspectors.
-        /// </summary>
-        [Obsolete("This API is obsolete and will be removed in the next version (Razor4 doesn't use CodeDom for code-generation)!")]
-        public IEnumerable<ICodeInspector> CodeInspectors
-        {
-            get { return _innerConfig.CodeInspectors; }
-        }
-#endif
-        
+                
         /// <summary>
         /// Gets the reference resolver.
         /// </summary>