+++ src/source/RazorEngine.Core/Configuration/ITemplateServiceConfiguration.cs
@@ -4,7 +4,6 @@
     using System.Collections.Generic;
 
     using Compilation;
-    using Compilation.Inspectors;
     using Templating;
     using Text;
     using RazorEngine.Compilation.ReferenceResolver;
@@ -29,14 +28,6 @@
         /// Gets the base template type.
         /// </summary>
         Type BaseTemplateType { get; }
-        
-#if !RAZOR4
-        /// <summary>
-        /// Gets the code inspectors.
-        /// </summary>
-        [Obsolete("This API is obsolete and will be removed in the next version (Razor4 doesn't use CodeDom for code-generation)!")]
-        IEnumerable<ICodeInspector> CodeInspectors { get; }
-#endif
 
         /// <summary>
         /// Gets the reference resolver.