+++ src/source/RazorEngine.Core/Legacy/Templating/TemplateService.cs
@@ -10,7 +10,6 @@ namespace RazorEngine.Templating
     using System.Reflection;
 
     using Compilation;
-    using Compilation.Inspectors;
     using Configuration;
     using Parallel;
     using Text;